\name{alpha.phi.ds}
\alias{alpha.phi.ds}
\title{Combines some estimates (alpha and phi) obtained from the different studies}
\usage{
  alpha.phi.ds(N, npara, M_study, alphaM, sum_p)
}
\arguments{
  \item{N}{}

  \item{npara}{}

  \item{M_study}{}

  \item{alphaM}{}

  \item{sum_p}{}
}
\value{
  a list that contains alpha and phi
}
\description{
  This function is required by the function
  \code{datashield.gee}
}

