\name{beta.vect.ds}
\alias{beta.vect.ds}
\title{Generates combined final estimates and standard errors of the fitted model}
\usage{
  beta.vect.ds(score, infomatrix, J.matrix, start.betas)
}
\arguments{
  \item{score}{}

  \item{infomatrix}{}

  \item{J.matrix}{}

  \item{start.betas}{}
}
\value{
  a list that contains the beta estimates and standard
  errors
}
\description{
  Function required by the function \code{datashield.gee}
}

