% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.coxph.R
\name{ds.coxph}
\alias{ds.coxph}
\title{Computes the statistical mean of a given vector}
\usage{
ds.coxph(data, survival_time, survival_event, terms, datasources = NULL)
}
\arguments{
\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}

\item{x}{a character, the name of a numerical vector}

\item{type}{a character which represents the type of analysis to carry out. 
If \code{type} is set to 'combine', a global mean is calculated 
if \code{type} is set to 'split', the mean is calculated separately for each study.}
}
\value{
a numeric
}
\description{
This function is similar to the R function \code{mean}.
}
\details{
It is a wrapper for the server side function.
}
\examples{
{

  # load that contains the login details
  data(logindata)

  # login and assign specific variable(s)
  myvar <- list('LAB_TSC')
  opals <- datashield.login(logins=logindata,assign=TRUE,variables=myvar)

  # Example 1: compute the pooled statistical mean of the variable 'LAB_TSC' - default behaviour
  ds.mean(x='D$LAB_TSC')

  # Example 2: compute the model
  ds.mean(x='D$LAB_TSC')

  # clear the Datashield R sessions and logout
  datashield.logout(opals)

}

}
\seealso{
\code{ds.quantileMean} to compute quantiles.

\code{ds.summary} to generate the summary of a variable.
}
\author{
Gaye A., Isaeva I.
}
