% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.coxph.R
\name{ds.coxph}
\alias{ds.coxph}
\title{Runs a cox proportional-hazards model.}
\usage{
ds.coxph(survival_time = NULL, survival_event = NULL, terms = NULL,
  data = NULL, maxit = 600, datasources = NULL)
}
\arguments{
\item{survival_time}{a character, the name of a numerical vector}

\item{survival_event}{a character, the name of a numerical vector}

\item{terms}{a character, the name of a numerical vector}

\item{data}{a character, the name of an optional data frame containing the variables in}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
a named vector of feature weights
}
\description{
A function that fits a Cox proportional hazards regression model.
}
\details{
It is a wrapper for the server side function.
}
\examples{
{

  # load that contains the login details
  data(logindata)

  # login
  opals <- datashield.login(logins=logindata,assign=TRUE)

  # Example 1: run a Cox proportional-hazards model
  beta <- ds.coxph(data = "D", survival_time = "D$TIME", survival_event = "D$CENSOR", 
                   terms = "D$AAE,D$BDS,D$HU,D$CU,D$IVDUPN,D$IVDURN,D$NPDT,D$RACE,D$TREAT,D$SITE", maxit = 600)

  # clear the Datashield R sessions and logout
  datashield.logout(opals)

}

}
\author{
Inberg, G
}
