% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.coxph.R
\name{ds.coxph}
\alias{ds.coxph}
\title{Runs the distributed Cox model learning algorithm.}
\usage{
ds.coxph(data = NULL, survival_time = NULL, survival_event = NULL,
  terms = NULL, maxit = 600, datasources = NULL)
}
\arguments{
\item{data}{a character, the name of the table that holds the data.}

\item{survival_time}{a character, the name of the survival time column.}

\item{survival_event}{a character, the name of a the survival event column.}

\item{terms}{a comma separated string containing the model features.}

\item{maxit}{numeric, max number of algorithm iterations.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{data frame}, from opal datasources.}
}
\value{
a named vector of feature weights
}
\description{
A function that fits a distributed Cox proportional hazards regression model.
}
\details{
It enables a distributed approach of the Cox proportional model. The advantage is that (sensitive) data
doesn't have to be uploaded to a central location for the calculations to be performed.
}
\examples{
{

  # load that contains the login details
  data(logindata)

  # login
  opals <- datashield.login(logins=logindata,assign=TRUE)

  # Example 1: run a Cox proportional-hazards model
  beta <- ds.coxph(data = "D", survival_time = "TIME", survival_event = "CENSOR", 
                   terms = "AAE,BDS,HU,CU,IVDUPN,IVDURN,NPDT,RACE,TREAT,SITE", maxit = 500)

  # clear the Datashield R sessions and logout
  datashield.logout(opals)

}

}
\author{
Inberg, G
}
